refactor(taskManager): 重构签到与投票管理路由代码
- 将任务管理路由中签到任务和投票任务子路由迁移至独立文件并抽象为单独模块
- 删除 TaskManagerRoute.kt 中大量具体处理逻辑，改为调用 checkInManagerRoute 和 voteManagerRoute
- 将签到任务相关路由逻辑迁移到 taskManagerRoutes/CheckInRoute.kt
- 优化导出 Excel 和用户干预操作逻辑，提升代码结构清晰度和可维护性
- 重命名子包 subroutes 为 subRoutes，统一包命名规范
- 删除旧模块中的序列化数据类和数据库查询辅助函数，相关功能已迁移
- 保持接口功能不变，重构后代码更加模块化和易于扩展
